% Auto-generated: do not edit by hand
\docType{package}
\name{dashPlayer-package}
\alias{dashPlayer}
\title{Interactive Media Player Component for Dash
}
\description{
Dash Player is a dash component for playing a variety of URLs, including file paths, YouTube, Facebook, Twitch, SoundCloud, Streamable, Vimeo, Wistia, Mixcloud, and DailyMotion. It is wrapped around the react-player component.

}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plot.ly>
}
